+++ addon/enums/i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-t-execution-variant.js
@@ -1,0 +1,8 @@
+import { createEnum } from 'ember-flexberry-data/utils/enum-functions';
+
+export default createEnum({
+  Unknown: 'Неизвестно',
+  Unexecuted: 'Не выполнено',
+  Executed: 'Выполнено',
+  Failed: 'Ошибка',
+});