+++ addon/enums/i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-t-type-of-audit-operation.js
@@ -1,0 +1,9 @@
+import { createEnum } from 'ember-flexberry-data/utils/enum-functions';
+
+export default createEnum({
+  INSERT: 'Создание',
+  UPDATE: 'Изменение',
+  DELETE: 'Удаление',
+  SELECT: 'Чтение',
+  EXECUTE: 'Выполнение',
+});