+++ addon/mixins/offline-model.js
@@ -4,6 +4,7 @@
 
 import Ember from 'ember';
 import DS from 'ember-data';
+import AuditModelMixin from './audit-model';
 
 /**
   Mixin for Ember models.
@@ -20,35 +21,7 @@ import DS from 'ember-data';
   @extends <a href="http://emberjs.com/api/classes/Ember.Mixin.html">Ember.Mixin</a>
   @public
 */
-export default Ember.Mixin.create({
-  /**
-    Creation date and time of model.
-
-    @property createTime
-    @type Date
-  */
-  createTime: DS.attr('date'),
-  /**
-    Name of user who created model.
-
-    @property creator
-    @type String
-  */
-  creator: DS.attr('string'),
-  /**
-    Date and time of last changes in model.
-
-    @property editTime
-    @type Date
-  */
-  editTime: DS.attr('date'),
-  /**
-    Name of user who changed model last time.
-
-    @property editor
-    @type String
-  */
-  editor: DS.attr('string'),
+export default Ember.Mixin.create(AuditModelMixin, {
   /**
     Date and time of last sync down of model.
 
@@ -56,6 +29,7 @@ export default Ember.Mixin.create({
     @type Date
   */
   syncDownTime: DS.attr('date'),
+
   /**
     Flag to indicate that model synchronized in readonly mode.
     Readonly mode allows to prevent any modifications of model on client side or server side.
@@ -65,12 +39,6 @@ export default Ember.Mixin.create({
     @readOnly
   */
   readOnly: DS.attr('boolean'),
-
-  /**
-  */
-  currentUserName: Ember.computed(function() {
-    return 'userName';
-  }).readOnly(),
 
   /**
     Global instance of Syncer that contains methods to sync model up and down.
@@ -91,8 +59,13 @@ export default Ember.Mixin.create({
     this.set('syncer', syncer);
   },
 
-  /*
-    Overrides saving method of model to set metadata properties.
+  /**
+    Save the record and persist any changes to the record to an external source via the adapter.
+    [More info](http://emberjs.com/api/data/classes/DS.Model.html#method_save).
+
+    @method save
+    @param {Object} [options]
+    @return {Promise}
   */
   save() {
     let _this = this;
@@ -101,21 +74,8 @@ export default Ember.Mixin.create({
       if (_this.get('readOnly')) {
         reject(new Error('Attempt to save readonly model instane.'));
       } else {
-        let currentDate = new Date();
-        let currentUser = _this.get('currentUserName');
-        let hasDirtyAttributes = _this.get('hasDirtyAttributes');
-        if (_this.get('isNew')) {
-          _this.set('createTime', currentDate);
-          _this.set('creator', currentUser);
-        }
-
-        if (hasDirtyAttributes && !_this.get('isDeleted')) {
-          _this.set('editTime', currentDate);
-          _this.set('editor', currentUser);
-        }
-
         let store = Ember.getOwner(_this).lookup('service:store');
-        if (hasDirtyAttributes && !store.get('offlineGlobals.isOnline')) {
+        if (_this.get('hasDirtyAttributes') && !store.get('offlineGlobals.isOnline')) {
           _this.get('syncer').createJob(_this).then((auditEntity) => {
             if (auditEntity.get('objectPrimaryKey')) {
               resolve(__super.call(_this, ...arguments));