+++ addon/mixins/regenerated/serializers/i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-audit-entity.js
@@ -1,0 +1,14 @@
+import Ember from 'ember';
+
+export let Serializer = Ember.Mixin.create({
+  attrs: {
+    user: { serialize: 'odata-id', deserialize: 'records' },
+    objectType: { serialize: 'odata-id', deserialize: 'records' },
+    auditFields: { serialize: false, deserialize: 'records' },
+  },
+
+  /**
+    Field name where object identifier is kept.
+  */
+  primaryKey: '__PrimaryKey',
+});