+++ addon/serializers/i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-audit-field-offline.js
@@ -3,6 +3,6 @@ import { Serializer } from 'ember-flexberry-data';
 
 export default Serializer.Offline.extend(DS.EmbeddedRecordsMixin, {
   attrs: {
-    ICSSoftSTORMNETBusinessAuditObjectsAuditEntity: { serialize: 'id', deserialize: 'records' },
+    auditEntity: { serialize: 'id', deserialize: 'records' },
   },
 });