+++ tests/unit/serializers/i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-audit-field-test.js
@@ -1,0 +1,14 @@
+import { moduleForModel, test } from 'ember-qunit';
+
+moduleForModel('i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-audit-field', 'Unit | Serializer | audit-field', {
+  needs: [
+    'serializer:i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-audit-field',
+    'model:i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-audit-entity',
+  ],
+});
+
+test('it serializes records', function(assert) {
+  let record = this.subject();
+  let serializedRecord = record.serialize();
+  assert.ok(serializedRecord);
+});