+++ tests/unit/serializers/i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-object-type-offline-test.js
@@ -5,14 +5,9 @@ import startApp from 'dummy/tests/helpers/start-app';
 
 let App;
 
-moduleForModel('i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-audit-entity', 'Unit | Serializer | audit-entity-offline', {
+moduleForModel('i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-object-type', 'Unit | Serializer | object-type-offline', {
   needs: [
-    'serializer:i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-audit-entity-offline',
-    'transform:i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-t-execution-variant',
-    'transform:i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-t-type-of-audit-operation',
-    'model:i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-audit-field',
-    'model:i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-object-type',
-    'model:i-c-s-soft-s-t-o-r-m-n-e-t-security-agent',
+    'serializer:i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-object-type-offline',
   ],
 
   beforeEach() {
@@ -27,7 +22,7 @@ moduleForModel('i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-audit-entity',
   },
 });
 
-test('this is a sure serializer', function(assert) {
+test('it serializes records', function(assert) {
   let record = this.subject(App.__container__.ownerInjection());
   let store = App.resolveRegistration('service:store').create(App.__container__.ownerInjection());
   let onlineSerializer = store.serializerFor(record._createSnapshot().modelName, true);